WIP: hash assets
try to move into a plugin to see if it helps order of
operations issue
