+++ src/css/_temp.scss
@@ -30,6 +30,7 @@ body {
 .inner h1 {
   font-size: var(--step-6-8);
   max-width: 10ch;
+  color: fuschia;
 }
 
 .inner p {
