+++ src/index.html
@@ -5,8 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <title>Breathe Easy Sheffield</title>
     <link id="asset-hash" href="/css/styles.css" rel="stylesheet" />
-    <!-- what happens if there's no file? -->
-    <script id="asset-hash" src="test1"></script>
   </head>
   <body>
     <div class="cont">
